Fix mypy errors and NameErrors in multiple modules
- Fix 132 errors in statest/statistics.py including comparison-overlap and signature mismatches
- Resolve NameErrors for MarginalProtocol, Callable, and LejaWeightingProtocol by adding missing imports
- Fix dtype mismatch in statest/groupacv tests for Torch backend
- Update util/linalg with robust None checks and proper typing
- Ensure all tests in pyapprox/statest pass on both NumPy and Torch backends

+++ pyapprox/statest/groupacv/tests/test_standalone.py
@@ -84,9 +84,9 @@ class TestGroupACVUtils:
         )
 
         # Check first few subsets
-        bkd.assert_allclose(subsets[0], bkd.asarray([0]))
-        bkd.assert_allclose(subsets[1], bkd.asarray([1]))
-        bkd.assert_allclose(subsets[2], bkd.asarray([2]))
+        bkd.assert_allclose(subsets[0], bkd.asarray([0], dtype=int))
+        bkd.assert_allclose(subsets[1], bkd.asarray([1], dtype=int))
+        bkd.assert_allclose(subsets[2], bkd.asarray([2], dtype=int))
 
     def test_allocation_matrix_is(self, bkd):
         """Test independent sampling allocation matrix."""
